pyproject.toml
[project]
name = "computer-use"
version = "0.1.0"
description = "Cross-platform computer use agent with multi-tier accuracy system"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "crewai[tools]>=0.86.0",
    "browser-use>=0.1.28",
    "pydantic>=2.0.0",
    "pillow>=10.0.0",
    "pyautogui>=0.9.54",
    "psutil>=5.9.0",
    "python-dotenv>=1.0.0",
    "opencv-python>=4.8.0",
    "paddlepaddle>=2.6.0",
    "paddleocr>=2.7.0",
    "easyocr>=1.7.0",
    "numpy>=1.24.0",
    "pyyaml>=6.0.0",
    "langchain-openai>=0.0.5",
    "langchain-anthropic>=0.1.0",
    "langchain-google-genai>=0.0.5",
    "langchain-community>=0.0.10",
    "rich>=14.2.0",
]

[project.optional-dependencies]
macos = [
    "atomacos>=0.5.0",
    "pyobjc-framework-ApplicationServices>=10.0",
    "pyobjc-framework-Cocoa>=10.0",
    "pyobjc-framework-Quartz>=10.0",
    "pyobjc-framework-Vision>=10.0",
]
windows = [
    "pywinauto>=0.6.8",
    "comtypes>=1.2.0",
]
linux = [
    "python3-pyatspi",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/computer_use"]

